requirements.txt
langchain
SentenceTransformer
NumPy
PyPDFLoader
Ollama